+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.61.0] - 2019-05-27
+### Added
+- Add `support` command, which allows logging into another account that has your support app installed.
+
+### Changed
+- Added funcionality to `browse` command to allow logging into an admin with a support token.
+
 ## [2.60.0] - 2019-05-25
 ### Added
 - Add `--qr` flag to the `browse` command, that outputs a QR code; intended for mobile devices.